requirements.txt
spacy
scikit-learn
pandas
numpy

tensorflow

kagglehub

langchain
langchain-core
langchain-community

transformers
huggingface_hub

fastapi
uvicorn

pydantic

python-dotenv

streamlit